% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMTmoments.R
\name{BMTmoments}
\alias{BMTmoments}
\alias{BMTmoment}
\alias{BMTmgf}
\alias{BMTchf}
\alias{mBMT}
\title{The BMT Distribution Moments, Moment-Generating Function and 
  Characteristic Function.}
\usage{
BMTmoment(p3, p4, type.p.3.4 = "t w", p1 = 0, p2 = 1,
  type.p.1.2 = "c-d", order, type = "standardised", method = "quadrature")

BMTmgf(s, p3, p4, type.p.3.4 = "t w", p1 = 0, p2 = 1,
  type.p.1.2 = "c-d")

BMTchf(s, p3, p4, type.p.3.4 = "t w", p1 = 0, p2 = 1,
  type.p.1.2 = "c-d")

mBMT(order, p3, p4, type.p.3.4, p1, p2, type.p.1.2)
}
\arguments{
\item{p3, p4}{tails weights (\eqn{\kappa_l} and \eqn{\kappa_r}) or 
asymmetry-steepness (\eqn{\zeta} and \eqn{\xi}) parameters of the BMT 
distribution.}

\item{type.p.3.4}{type of parametrization asociated to p3 and p4. "t w" means
tails weights parametrization (default) and "a-s" means asymmetry-steepness
parametrization.}

\item{p1, p2}{domain (minimum and maximum) or location-scale (mean and 
standard deviation) parameters of the BMT ditribution.}

\item{type.p.1.2}{type of parametrization asociated to p1 and p2. "c-d" means
domain parametrization (default) and "l-s" means location-scale 
parametrization.}

\item{order}{order of the moment.}

\item{type}{type of the moment: raw, central or standardised (default).}

\item{method}{method to obtain the moment: exact formula or Chebyshev-Gauss 
quadrature (default).}

\item{s}{variable for the moment-generating and characteristic functions.}
}
\value{
\code{BMTmoment} gives any raw, central or standarised moment, 
  \code{BMTmgf} the moment-generating function and \code{BMTchf} the 
  characteristic function
  
  The arguments are recycled to the length of the result. Only the first 
  elements of \code{type.p.3.4}, \code{type.p.1.2}, \code{type} and
  \code{method} are used.
  
  If \code{type.p.3.4 == "t w"}, \code{p3 < 0} and \code{p3 > 1} are errors 
  and return \code{NaN}.
  
  If \code{type.p.3.4 == "a-s"}, \code{p3 < -1} and \code{p3 > 1} are errors 
  and return \code{NaN}.
  
  \code{p4 < 0} and \code{p4 > 1} are errors and return \code{NaN}.
  
  If \code{type.p.1.2 == "c-d"}, \code{p1 >= p2} is an error and returns 
  \code{NaN}.
  
  If \code{type.p.1.2 == "l-s"}, \code{p2 <= 0} is an error and returns 
  \code{NaN}.
}
\description{
Any raw, central or standarised moment, the moment-generating 
  function and the characteristic function for the BMT distribution, with 
  \code{p3} and \code{p4} tails weights (\eqn{\kappa_l} and \eqn{\kappa_r}) 
  or asymmetry-steepness parameters (\eqn{\zeta} and \eqn{\xi}) and \code{p1}
  and \code{p2} domain (minimum and maximum) or location-scale (mean and 
  standard deviation) parameters.
}
\details{
See References.
}
\examples{
layout(matrix(1:4, 2, 2, TRUE))
s <- seq(-1, 1, length.out = 100)

# BMT on [0,1] with left tail weight equal to 0.25 and 
# right tail weight equal to 0.75
BMTmoment(0.25, 0.75, order = 5) # hyperskewness by Gauss-Legendre quadrature
BMTmoment(0.25, 0.75, order = 5, method = "exact") # hyperskewness by exact formula
mgf <- BMTmgf(s, 0.25, 0.75) # moment-generation function
plot(s, mgf, type="l")
chf <- BMTchf(s, 0.25, 0.75) # characteristic function

# BMT on [0,1] with asymmetry coefficient equal to 0.5 and 
# steepness coefficient equal to 0.5
BMTmoment(0.5, 0.5, "a-s", order = 5)
BMTmoment(0.5, 0.5, "a-s", order = 5, method = "exact")
mgf <- BMTmgf(s, 0.5, 0.5, "a-s")
plot(s, mgf, type="l")
chf <- BMTchf(s, 0.5, 0.5, "a-s")

# BMT on [-1.783489, 3.312195] with 
# left tail weight equal to 0.25 and 
# right tail weight equal to 0.75
BMTmoment(0.25, 0.75, "t w", -1.783489, 3.312195, "c-d", order = 5)
BMTmoment(0.25, 0.75, "t w", -1.783489, 3.312195, "c-d", order = 5, method = "exact")
mgf <- BMTmgf(s, 0.25, 0.75, "t w", -1.783489, 3.312195, "c-d")
plot(s, mgf, type="l")
chf <- BMTchf(s, 0.25, 0.75, "t w", -1.783489, 3.312195, "c-d")

# BMT with mean equal to 0, standard deviation equal to 1, 
# asymmetry coefficient equal to 0.5 and 
# steepness coefficient equal to 0.5
BMTmoment(0.5, 0.5, "a-s", 0, 1, "l-s", order = 5)
BMTmoment(0.5, 0.5, "a-s", 0, 1, "l-s", order = 5, method = "exact")
mgf <- BMTmgf(s, 0.5, 0.5, "a-s", 0, 1, "l-s")
plot(s, mgf, type="l")
chf <- BMTchf(s, 0.5, 0.5, "a-s", 0, 1, "l-s")
}
\references{
Torres-Jimenez, C. J. and Montenegro-Diaz, A. M. (2017, September), 
  \emph{An alternative to continuous univariate distributions supported on a 
  bounded interval: The BMT distribution}. ArXiv e-prints.
  
  Torres-Jimenez, C. J. (2018), \emph{The BMT Item Response Theory model: A 
  new skewed distribution family with bounded domain and an IRT model based 
  on it}, PhD thesis, Doctorado en ciencias - Estadistica, Universidad 
  Nacional de Colombia, Sede Bogota.
}
\seealso{
\code{\link{BMTcentral}}, \code{\link{BMTdispersion}}, 
  \code{\link{BMTskewness}}, \code{\link{BMTkurtosis}} for specific 
  descriptive measures or moments.
}
\author{
Camilo Jose Torres-Jimenez [aut,cre] \email{cjtorresj@unal.edu.co}
}
